+++ apps/codeforafrica/src/components/NavList/index.test.js
@@ -1,8 +1,12 @@
+import { createRender } from "@commons-ui/testing-library";
 import React from "react";
 
 import NavList from ".";
 
-import { render } from "@/codeforafrica/utils/test";
+import theme from "@/codeforafrica/theme";
+
+// eslint-disable-next-line testing-library/render-result-naming-convention
+const render = createRender({ theme });
 
 describe("<NavList />", () => {
   it("renders unchanged", () => {
