--- apps/codeforafrica/src/components/NavMenu/index.snap.js
@@ -1,61 +1,0 @@
-// Jest Snapshot v1, https://goo.gl/fbAQLP
-
-exports[`<NavMenu /> renders unchanged 1`] = `
-<div>
-  <ul
-    class="css-69dg8j"
-    ownerstate="[object Object]"
-  >
-    <li
-      class="css-17g3syn"
-    >
-      <a
-        class="css-nmd6m8"
-        href="/our-work"
-      >
-        Our Work
-      </a>
-    </li>
-    <li
-      class="css-17g3syn"
-    >
-      <a
-        class="css-nmd6m8"
-        href="/about"
-      >
-        About
-      </a>
-    </li>
-    <li
-      class="css-17g3syn"
-    >
-      <a
-        class="css-nmd6m8"
-        href="/about"
-      >
-        Stories
-      </a>
-    </li>
-    <li
-      class="css-17g3syn"
-    >
-      <a
-        class="css-nmd6m8"
-        href="/opportunity"
-      >
-        Opportunity
-      </a>
-    </li>
-    <li
-      class="css-17g3syn"
-    >
-      <a
-        class="css-nmd6m8"
-        href="/contact"
-      >
-        Contact
-      </a>
-    </li>
-  </ul>
-</div>
-`;
